Feature/sg 1106 Minor fixes in export api (#1412)
* Fix wrong type annotation for quantization_mode

* Fix initialization of example_input_image

* Added provides to stay compatible with newer onnxruntime that require provides to be present

* Fixed edge case of exporting detection models of small size and num_pre_nms_predictions that exceeds this value

* Change initilization of grid in yolox/yolonas/ppyoloe to ensure that constants in ONNX file is fp32 and not fp64.
https://github.com/Deci-AI/super-gradients/issues/1392

+++ src/super_gradients/training/models/detection_models/yolo_nas/dfl_heads.py
@@ -281,14 +281,16 @@ class NDFLHeads(BaseDetectionModule, SupportsReplaceNumClasses):
             else:
                 h = int(self.eval_size[0] / stride)
                 w = int(self.eval_size[1] / stride)
-            shift_x = torch.arange(end=w) + self.grid_cell_offset
-            shift_y = torch.arange(end=h) + self.grid_cell_offset
+
+            shift_x = torch.arange(end=w, dtype=dtype) + self.grid_cell_offset
+            shift_y = torch.arange(end=h, dtype=dtype) + self.grid_cell_offset
+
             if torch_version_is_greater_or_equal(1, 10):
                 shift_y, shift_x = torch.meshgrid(shift_y, shift_x, indexing="ij")
             else:
                 shift_y, shift_x = torch.meshgrid(shift_y, shift_x)
 
-            anchor_point = torch.stack([shift_x, shift_y], dim=-1).to(dtype=dtype)
+            anchor_point = torch.stack([shift_x, shift_y], dim=-1)
             anchor_points.append(anchor_point.reshape([-1, 2]))
             stride_tensor.append(torch.full([h * w, 1], stride, dtype=dtype))
         anchor_points = torch.cat(anchor_points)
